requirements.txt
# Procesamiento de audio
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Transcripción
openai-whisper>=20231117

# Speaker Diarization
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0

# Utilidades
tqdm>=4.66.0
numpy>=1.24.0

# Opcional: para procesamiento de audio avanzado
ffmpeg-python>=0.2.0

